Add Gitbook Action to publish HTML, PDF, EPUB, and MOBI
Add Gitbook Action to publish this document in the
gh-pages branch of the same repository in HTML, PDF,
EPUB, and MOBI formats.  The draft and release
versions of the documents are published in different
directories of the gh-pages branch so all published
versions are available for download and online
viewing.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>